Move Builders as non inner classes in sdk-extension/logging
